+++ R/print.xtable.fdt.quantile.R
@@ -1,0 +1,8 @@
+print.xtable.fdt.quantile <- function(x,
+                                      ...,
+                                      include.rownames = FALSE)
+{
+  NextMethod("print",
+             include.rownames = include.rownames,
+             ...)
+}