+++ R/print.xtableList.fdt.quantile.multiple.R
@@ -1,0 +1,8 @@
+print.xtableList.fdt.quantile.multiple <- function(x,
+                                                   ...,
+                                                   include.rownames = FALSE)
+{
+  NextMethod("print",
+             include.rownames = include.rownames,
+             ...)
+}